+++ CHANGELOG.md
@@ -6,3 +6,9 @@
 
 * 0.1.1: bug fix
     - remove unnecessary imports and ChangeNotifier
+
+* 0.1.2: bug fix; better docs
+    - deal with shonky iOS now playing information, which variously returns
+      images and persistent IDs as empty or zero
+    - improve in-code documentation
+